\name{DataAll4m-package}
\alias{DataAll4m-package}
\alias{DataAll4m}
\docType{package}
\title{DataAll4m: Analiza danych sprzedazowych sklepu muzycznego All4m}

\description{
Pakiet DataAll4m powstal do ulatwienia analizy danych sprzedazowych na Allegro sklepu muzycznego All4m.
Funkcje dostepne w pakiecie tworza tabele i wykresy przedstawiajace dane o sprzedazy naszej oraz konkurencji,
biorac pod uwage podzial na kategorie, promowanie, itp.
}

\details{
Pakiet DataAll4m ma na celu przetwarzanie danych Allegro do wygodnych do analizy postaci:
\itemize{
\item ranking w wybranej kategorii w formie wykresu slupkowego, z rozroznieniem akcji promowanych
i niepromowanych za pomoca kolorow;
\item wartosc sprzedazy w wybranej kategorii z podzialem na klasy cenowe w formie wykresu slupkowego,
z rozroznieniem aukcji promowanych i niepromowanych za pomoca kolorow;
\item tabele przedstawiajace wartosci w poszczegolnych kategoriach, nasze oraz ogolne
\item tabela przedstawiajaca ranking w wybranej kategorii
}
}

\author{Sklep Muzyczny all4m - Karolina Naskret}