% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelSzt.R
\name{modelSzt}
\alias{modelSzt}
\title{Model opisujący liczbę sprzedanych sztuk na osobę}
\usage{
modelSzt(cena)
}
\arguments{
\item{cena}{double - cena produktu.}
}
\description{
Funkcja modelSzt() dla podanej ceny oblicza przybliżoną liczbę sprzedanych sztuk na osobę.
Funkcja działa dla pojedynczej wartości, sama w sobie jest mało przydatna. Jest jedynie potrzebna
do działania funkcji vmodelSzt(), która jest zwektoryzowaną wersją tej funkcji, i to jej polecam
używać.
}
\examples{
modelSzt(35)

}
