% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/napiszKat.R
\name{napiszKat}
\alias{napiszKat}
\title{Napis przedstawiajay kategorie}
\usage{
napiszKat(id_kat, symbol = " -> ")
}
\arguments{
\item{symbol}{symbol łączący kolejne gałęzie kategorii}

\item{nazwa}{id_kat integer - id interesujacej nas kategorii}
}
\description{
Funkcja napiszKat() przeszukuje baze kategorii Allegro i wypisuje wszystkie kategorie i podkategorie,
polaczone znakiem strzalki
}
\examples{
napiszKat(67307)
napiszKat(67307, symbol = "-")

}
