% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naszeWartosci.R
\name{naszeWartosci}
\alias{naszeWartosci}
\title{Wartosci naszej sprzedazy dla wybranych kategorii}
\usage{
naszeWartosci(dane, kategorie = unique(dane$Id_kategorii),
  group = FALSE, id = 169890, info_promowania = TRUE)
}
\arguments{
\item{dane}{dane przygotowane przez funkcje przygotujDane();}

\item{kategorie}{wektor interesujących nas kategorii. Domyslnie sa wziete wszystkie mozliwe kategorie
w bazie dnaych dane}

\item{group}{wartosc logiczna okreslajaca, czy chcemy podane wczesniej kategorie scalic w jedna 
szersza kategorie}

\item{id}{liczba bedaca identyfikatorem interesujacego nas uzytkownika Allegro.
Domyslnie id = 169890 (czyli my)}

\item{info_promowanie}{wartosc logiczna okreslajaca, czy chcemy wyswietlic udzial aukcji promowanych}
}
\description{
Funkcja naszeWartości() tworzy tabele, w ktorej dla kazdej interesujacej nas kategorii Allegro podana 
jest wartosc naszej sprzedazy. Podaje rowniez liczbe naszych aukcji promowanych oraz ich wartosc.
Jest rowniez mozliwe obliczenie naszej wartosci dla szerszej kategorii, np. "Instrumenty".
}
\examples{
naszeWartosci(aukcje)
naszeWartosci(aukcje, kategorie = c(122815, 122816, 122818, 122819), group = TRUE)
naszeWartosci(aukcje, kategorie = 122815, info_promowania = FALSE)

}
