% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/przygotujDaneCSV.R
\name{przygotujDaneCSV}
\alias{przygotujDaneCSV}
\title{Pobranie i przygotowanie danych z pliku csv}
\usage{
przygotujDaneCSV(adres_aktywnych, adres_zakonczonych, model = TRUE)
}
\arguments{
\item{adres_aktywnych}{sciezka do pliku w formacie csv.}

\item{adres_zakonczonych}{sciezka do pliku w formacie csv.}

\item{model}{wartość logiczna określająca, czy funkcja ma użyć modelu do przybliżenia liczby sprzedanych sztuk}
}
\description{
Funkcja przygotujDaneCSV() sczytuje dane, osobno aukji aktywnych i zakończonych (format csv) 
oraz wstepnie przygotowuje je do pracy z kolejnymi funkcjami w tym pakiecie.
}
\examples{
przygotujDaneCSV(".../active.csv", ".../closed.csv")
przygotujDaneCSV(".../active169890.csv", ".../closed169890.csv", model = FALSE)

}
