% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/przygotujDane.R
\name{przygotujDaneXLSX}
\alias{przygotujDaneXLSX}
\title{Pobranie i przygotowanie danych z pliku xlsx}
\usage{
przygotujDaneXLSX(adres_danych)
}
\arguments{
\item{adres_danych}{sciezka do pliku w formacie xlsx.}
}
\description{
Funkcja przygotujDaneXLSX() sczytuje dane (format xlsx) oraz wstepnie przygotowuje je do pracy 
z kolejnymi funkcjami w tym pakiecie.
}
\examples{
przygotujDaneXLSX(".../aukcje.xlsx")

}
