% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/przygotujDoWykresu.R
\name{przygotujDoWykresu}
\alias{przygotujDoWykresu}
\title{Przygotowanie danych do narysowania z nich wykresu}
\usage{
przygotujDoWykresu(dane, kategoria, k = 30, id = 169890)
}
\arguments{
\item{dane}{dane przygotowane przez funkcje przygotujDane();}

\item{kategoria}{liczba (wektor) bedaca id interesujacej nasz kategorii}

\item{k}{liczba osob, ktore chcemy zawrzec na wykresie; domyslnie k = 30}

\item{id}{id sprzedawcy, ktorego chcemy wyroznic; domyslnie nasze id, czyli 169890}
}
\description{
Funkcja przygotujDoWykresu() sama w sobie nie robi nic ciekawego. Ma ona za zadanie przetworzyc
dane tak, zeby potem wygodnie bylo z niej rysowac wykres obrazujacy ranking.
}
\examples{
przygotujDoWykresu(aukcje, 122815)

}
