% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rysujRanking.R
\name{rysujRanking}
\alias{rysujRanking}
\title{Wykres przedstawiajacy ranking sprzedawcow w konkretnej kategorii}
\usage{
rysujRanking(dane, kategoria, k = 30, id = 169890)
}
\arguments{
\item{dane}{dane przygotowane przez funkcje przygotujDane();}

\item{kategoria}{liczba (wektor) bedaca id interesujacej nasz kategorii}

\item{k}{liczba osob, ktore chcemy zawrzec w rankingu; domyslnie k = 30}

\item{id}{id sprzedawcy, ktorego chcemy wyroznic; domyslnie nasze id, czyli 169890}
}
\description{
Funkcja rysujRanking() tworzy wykres, na ktorym dla interesujacej nas kategorii Allegro przedstawione
sa wartosci dla pierwszych k sprzedawcow oraz udzial aukcji promowanych oraz niepromowanych.
Slupek przedstawiajacy nasza wartosc wyroznia sie ciemniejszymi kolorami.
Jest rowniez mozliwe narysowanie rankingu dla szerszej kategorii, np. "Instrumenty".
}
\examples{
rysujRanking(aukcje, 122815)
rysujRanking(aukcje, kategoria = c(122815, 122816, 122818, 122819))
rysujRanking(aukcje, 122819, k = 10)

}
