% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sczytajSztuki.R
\name{sczytajSztuki}
\alias{sczytajSztuki}
\title{Pobranie danych o liczbie sprzedanych sztuk z aukcji Allegro}
\usage{
sczytajSztuki(nr_aukcji)
}
\arguments{
\item{nr_aukcji}{numer interesujacej nas aukcji.}
}
\description{
Funkcja sczytajSztuki() sczytuje ze strony aukci Allegro liczbe osob, ktore kupily ten produkt
oraz liczbe sprzedanych sztuk.
}
\examples{
sczytajSztuki(7387674525)

}
