% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stworzRanking.R
\name{stworzRanking}
\alias{stworzRanking}
\title{Ranking w podanej kategorii Allegro}
\usage{
stworzRanking(dane, kategoria, k = 30)
}
\arguments{
\item{dane}{dane przygotowane przez funkcje przygotujDane();}

\item{kategoria}{liczba (wektor) bedaca id interesujacej nasz kategorii}

\item{k}{liczba osob, ktore chcemy zawrzec w rankingu; domyslnie k = 30}
}
\description{
Funkcja stworzRanking() tworzy tabele, w ktorej dla interesujacej nas kategorii Allegro podane jest
pierwszych k sprzedawcow o najwiekszej wartosci z informacja o ich wartosci, udziale aukcji
promowanych oraz niepromowanych. Jest rowniez mozliwe stworzenie rankingu
dla szerszej kategorii, np. "Instrumenty".
}
\examples{
stworzRanking(aukcje, 122815)
naszeWartosci(aukcje, kategoria = c(122815, 122816, 122818, 122819))
naszeWartosci(aukcje, 122819, k = 10)

}
