% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wartoscKategorii.R
\name{wartoscKategorii}
\alias{wartoscKategorii}
\title{Wartosc ogolna wybranej kategorii}
\usage{
wartoscKategorii(dane, kategorie = unique(dane$Id_kategorii),
  group = FALSE)
}
\arguments{
\item{dane}{dane przygotowane przez funkcje przygotujDane();}

\item{kategorie}{wektor interesujacych nas kategorii. Domyslnie sa wziete wszystkie mozliwe kategorie
w bazie dnaych dane}

\item{group}{wartosc logiczna okreslajaca, czy chcemy podane wczesniej kategorie scalic w jedna 
szersza kategorie. Domyslnie group = FALSE}
}
\description{
Funkcja wartoscKatgeorii() tworzy tabele, w ktorej dla interesujacej nas kategorii Allegro podana 
jest wartosc ogolna tej kategorii. Podana jest rowniez wartosc aukcji promowanych i zwyklych.
Jest rowniez mozliwe scalenie kilku kategorii do jednej szerszej kategorii, np. "Instrumenty".
}
\examples{
wartoscKategorii(aukcje)
wartoscKategorii(aukcje, kategorie = c(122815, 122816, 122818, 122819), group = TRUE)
wartoscKategorii(aukcje, 122815)

}
