% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wykresKategorii.R
\name{wykresKategorii}
\alias{wykresKategorii}
\title{Wartości kategorii Allegro}
\usage{
wykresKategorii(dane, id = 169890)
}
\arguments{
\item{dane}{dane przygotowane przez funkcje przygotujDane();}

\item{id}{id sprzedawcy, ktorego chcemy wyroznic; domyslnie nasze id, czyli 169890}
}
\description{
Funkcja wykresKategorii() wykresy, na których dla każdej kategorii narysowana jest jej wartość,
z podziałem na aukcje promowane, niepromowane i nasze. Kategorie są posortowane według wartości
aukcji niepromowanych. Liczby na słupkach to średnie ceny aukcji w danej kategorii z tym samym podziałem.
}
\examples{
wk <- wykresKategorii(dane)
wk$wykres1

}
