% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wykresPodzialCen.R
\name{wykresPodzialCen}
\alias{wykresPodzialCen}
\title{Wykres przedstawiajacy sprzedaz z podzialem na klasy cenowe}
\usage{
wykresPodzialCen(dane, kategoria, podzial, typ,
  zakres = c(min(dane$Cena), max(dane$Cena)), s = 12, a = 90)
}
\arguments{
\item{dane}{dane przygotowane przez funkcje przygotujDane();}

\item{kategoria}{liczba (wektor) bedaca id interesujacej nasz kategorii}

\item{podzial}{wektor liczb bedacych granicami poszczegolnych kategorii}

\item{typ}{zmienna okreslajaca czy chcemy na wykresie przedstawic wartosc ("W"), czy ilosc ("I")}

\item{zakres}{zakres cenowy nas interesujacy}
}
\description{
Funkcja wykresPodzialCen() tworzy wykres, na ktorym dla interesujacej nas kategorii Allegro przedstawione
sa wartosci lub ilosci sprzedazy w poszczegolnych klasach cenowych.
Sposob podzialu na klasy cenowe okreslamy sami.
}
\examples{
WykresPodzialCen(aukcje, 260310, p, typ = "W", zakres= c(0, 2000))
WykresPodzialCen(aukcje, 260310, p, typ = "I")

}
