% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/znajdzIdKat.R
\name{znajdzIdKat}
\alias{znajdzIdKat}
\title{Id interesujacej nas kategorii Allegro}
\usage{
znajdzIdKat(nazwa, nr_podkat)
}
\arguments{
\item{nazwa}{string - nazwa (pod)kategorii ktora nas interesuje, np "Instrumenty", "Bebny i bebenki"
(bez przecinkow i polskich znakow);}

\item{nr_podkat}{integer - w drzewie kategorii Allegro mamy 7 mozliwych poziomow. nr_podkat
okresla poziom, ktory chcemy przeszukac}
}
\description{
Funkcja znajdzIdKat() przeszukuje baze kategorieAllegro i wypisuje wszystkie id (pod)kategorii, ktore
mieszcza sie w interesujacej nas kategorii
}
\examples{
znajdzIdKat("Instrumenty perkusyjne", 3)
znajdzIdKat("Kapodastry", 5)

}
